post-service/requirements.txt
grpcio-tools
sqlalchemy
psycopg2-binary

